(enigma2) fixed bug with svn number with not english systems (enigma2) added halt on compile error on make scripts (common) fixed failure on compilation of crossepg_xmltv
git-svn-id: http://crossepg.googlecode.com/svn/trunk@91 bda989b6-6649-11de-9c70-efa4e9705d34

+++ make_e2_oe1.5.sh
@@ -4,6 +4,10 @@ CFLAGS+=-I/opt/dmm/dm800/build/tmp/staging/mipsel-oe-linux/usr/include/libxml2 \
 CC=${CROSS}gcc \
 STRIP=${CROSS}strip \
 make
+if [ $? != 0 ]; then
+	echo compile error
+	exit 1
+fi
 
 [ -d tmp ] && rm -rf tmp
